==> src/10x/02_json_to_sf.Rmd <==
# Convert GeoJSON files into SF Objects
```{r}
# load libraries
library(sf) # for reading in spatial data
library(tidyverse) # for manipulating data frames
library(ggmap) # for getting map backgrounds
library(dotenv) # for environment variables
```

## Read & Clean Data
```{r}
# load data
p13u <- st_read(
    "/Users/ilyonsg/Documents/nrel/data/sfo/solar_extreme_batteries_high_timeseries_combined/P13U.json",
    drivers = "GeoJSON",
    stringsAsFactors = TRUE
)

# copy dataframe for cleaning
p13u_sf <- p13u
glimpse(p13u_sf)


# convert type column to factor
p13u_sf <- p13u_sf %>%
    mutate(
        name = as.character(name),
        length..km. = as.numeric(length..km.),
        ampacity = as.numeric(ampacity),
        inverter_size = as.numeric(inverter_size),
        kva = as.numeric(kva),
        kvar = as.numeric(kvar),
        kw_rated = as.numeric(kw_rated),
        kwh_rated = as.numeric(kwh_rated),
        array_size = as.numeric(array_size)
    )

glimpse(p13u_sf)
p13u_sf %>% sapply(levels)

# export sf object to RDS
saveRDS(p13u_sf, "data/10x/p13u_sf.rds")
```
Rows: 882,412
Columns: 24
$ is_switch     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ is_open       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ is_fuse       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ kva           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ kvar          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ kw_rated      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ kwh_rated     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ length..km.   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ ampacity      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ peak_voltage  <dbl> 0.9789937, 0.9798307, 0.9802534, 0.9706192, 0.9719708, 0…
$ rated_kv      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ array_size    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ inverter_size <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ control_type  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ max_kw        <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ max_kvar      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ is_ct         <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
$ type          <chr> "Node", "Node", "Node", "Node", "Node", "Node", "Node", …
$ name          <chr> "p13udm3166", "p13udm3168", "p13udm3169", "p13udm3170", …
$ marker.color  <chr> "000000", "000000", "000000", "000000", "000000", "00000…
$ marker.size   <chr> "small", "small", "small", "small", "small", "small", "s…
$ kvs           <list> <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>,…
$ phases        <list> <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>, <>,…
$ geometry      <POINT [°]> POINT (-122.2683 37.86478), POINT (-122.2682 37.86…

$type
[1] "Capacitor"   "Line"        "Load"        "Node"        "PVSystem"   
[6] "Storage"     "Transformer"

## Filter Data to Buses and Lines
```{r}
# look at the unique values of the type column
# read in RDS
p13u <- readRDS("data/10x/p13u_sf.rds")

# look at the names of the type "Load"
p13u %>%
    filter(type == "Storage") %>%
    select(name) %>%
    head()
# Loads
#   load_p13ulv2240
#   load_p13ulv86673
#   129330 loads
# Capacitors
#   p13uc0
#   p13uc48
#   count 178
# PVSystem
#   pv_load_p13ulv43357
#   pv_load_p13ulv68747
#   110023
# Storage
#   storage_p13ulv43357
#   storage_p13ulv11021
#   51621



# create lines.sf
lines.sf <- p13u %>%
    filter(type == "Line" | type == "Transformer")
lines.sf <- lines.sf %>%
    select(c(
        "is_switch", # lines
        "is_open", # lines
        "is_fuse", # fuse
        "kva", # transformer
        "type", # all
        "length..km.", # lines
        "ampacity", # lines
        "name", # all
        "kvs", # transformer
        "phases", # lines
        "geometry" # all
    )) 

# create buses.sf
buses.sf <- p13u %>%
    filter(type == "Node")
# drop columns that only contain NA
buses.sf <- buses.sf %>%
   select(c(
         "name",
         "peak_voltage",
         "geometry"
   ))

# rename peak voltage to voltage
buses.sf <- buses.sf %>%
    rename(voltage = peak_voltage)

# remove buses with duplicate geometry
buses.sf <- buses.sf %>%
    distinct(geometry, .keep_all = TRUE)

saveRDS(lines.sf, "data/10x/lines_sf.rds")
saveRDS(buses.sf, "data/10x/buses_sf.rds")
```
