==> src/10x/03_map.Rmd <==
# Setup Plotting with a Map
```{r}
library(ggmap) # for getting map backgrounds
library(dotenv) # for environment variables

#  define bounding box
bottom <- 37.7284
top <- 37.90437
left <- -122.3414
right <- -122.1507

# read in env file
load_dot_env(".env")
stadia.maps.api.key <- Sys.getenv("STADIA_MAPS_API_KEY")

# configure ggmap with api keys
register_stadiamaps(stadia.maps.api.key)

# get background map
map <- ggmap::get_stadiamap(
  bbox = c(
    left = left, 
    right = right, 
    bottom = bottom, 
    top = top
  ), 
  zoom = 13, 
  maptype = "stamen_terrain_lines"
)

# display map
ggmap(map)

# export map as RDS
saveRDS(map, "data/10x/map_stadia_z13.rds")
```
