==> src/10x/07_boundary.Rmd <==
# Boundary
```{r}
library(tidyverse)
library(sf)
library(ggmap)
library(terra)
library(h3jsr)
library(concaveman)
```

## Read in Data
```{r}
# read in RDS
lines.sf <- readRDS("data/10x/lines_sf.rds")
buses.sf <- readRDS("data/10x/buses_sf.rds")
map <- readRDS("data/10x/map_stadia_z14.rds")

# get the bounding box of map
left <- attr(map, "bb")$ll.lon
right <- attr(map, "bb")$ur.lon
bottom <- attr(map, "bb")$ll.lat
top <- attr(map, "bb")$ur.lat
```

## Read in Hexes
```{r}
# read in hexes
buses.h3 <- readRDS("data/10x/buses_h3_r8to12.rds")

# keep just resolution 9 hexes
buses.h3 <- buses.h3 %>% filter(res == 9) 

boundary <- cells_to_multipolygon(h3_addresses = buses.h3$h3, simple = TRUE)
plot(boundary)

# convert to SpatVector
boundary.sv <- terra::vect(boundary)

# export boundary to RDS
saveRDS(boundary.sv, "data/10x/boundary_r10_sv.rds")
```

## Convex Hull
```{r}
# convert buses.sf to 'SpatVector'
buses.sv <- terra::vect(buses.sf)
saveRDS(buses.sv, "data/10x/buses_sv.rds")

# create a convex hull around the buses
bus.convex.hull <- convHull(buses.sv, by="")

plot(bus.convex.hull)
# export
saveRDS(bus.convex.hull, "data/10x/buses_convex_hull.rds")
```

## Concave Hull
```{r}
# create a concave hull around the buses
bus.concave.c1 <- concaveman(buses.sf, concavity = 1)
bus.concave.c2 <- concaveman(buses.sf, concavity = 2)
bus.concave.c3 <- concaveman(buses.sf, concavity = 3)
bus.concave.c4 <- concaveman(buses.sf, concavity = 4)

plot(bus.concave.c1) # too detailed
plot(bus.concave.c2) # good
plot(bus.concave.c3) # decent, 
plot(bus.concave.c4) # too simple

# export 2 and 3
# convert to SpatVector
bus.concave.c1.sv <- terra::vect(bus.concave.c1)
bus.concave.c2.sv <- terra::vect(bus.concave.c2)
bus.concave.c3.sv <- terra::vect(bus.concave.c3)
saveRDS(bus.concave.c1.sv, "data/10x/buses_concave_c1_sv.rds")
saveRDS(bus.concave.c2.sv, "data/10x/buses_concave_c2_sv.rds")
saveRDS(bus.concave.c3.sv, "data/10x/buses_concave_c3_sv.rds")
```


# Other Methods with Buffering

## Create a Buffer Around Points
```{r}
# convert buses to 'SpatVector'
buses.sv <- terra::vect(buses.sf)

# create a 1000m buffer using terra 
buses.buffer.sv <- terra::buffer(buses.sv, 1000)

# export bus buffer to RDS
saveRDS(buses.buffer.sv, "data/10x/buses_buffer_1000m_vec.rds")
```


## Create Buffer Around Lines
```{r}
# merge the lines spatially into a single geometry
lines.all <- st_union(lines.sf)

# create a 50m buffer around lines.all
lines.buffer <- st_buffer(
    lines.all, 
    endCapStyle = "ROUND", 
    joinStyle = "ROUND",
    dist = 1000,
    # allow_holes = FALSE
    # max_cells = 10000,
)

# visualize the line buffer
plot(lines.buffer)

# export the line buffer to RDS
saveRDS(lines.buffer, "data/10x/lines_buffer_1000m.rds")
```
